🪛[chore]:  파일 분리 #2
